feat: 面接練習に深掘り（面接官の追い質問）を追加
4択クイズと同じ深掘り「なぜ?」チェーンを面接練習にも導入する。

- 深掘りチェーン FollowUpChain を共通コンポーネント（components/
  FollowUpChain.tsx）に切り出し、QuizView と InterviewView で共用
- InterviewQuestion 型に followUps（{q,a}[]）を追加
- 面接練習の模範回答内（メモの下・自己評価の上）に段階開示で表示
- 同梱の面接質問 35問すべてに面接官の追い質問を2つずつ作成
  （Q=面接官の追い質問／A=答え方の要点・未経験者向け）

Closes #549
